% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{colDistance}
\alias{colDistance}
\title{Calculate distances between columns in a matrix}
\usage{
colDistance(x, f = euclidean_distance)
}
\arguments{
\item{x}{Matrix with multiple columns}

\item{f}{A function to use for calculating distances}

\item{...}{Additional parameters needed for f}
}
\value{
A dist object representing distances between columns.
}
\description{
Calculate distances between columns of a matrix using the given
distance metric (as a function) and return a dist object.
}
