% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Extended Distance Matrix Computation}
\usage{
distance(
  x,
  method = c("euclidean", "pearson", "pearsonm", "rmsd", "multimetric", "maximum",
    "manhattan", "canberra", "binary", "minkowski", "presence"),
  diag = FALSE,
  upper = FALSE,
  p = 2,
  which = c("row", "column"),
  ...
)
}
\arguments{
\item{p}{}
}
\description{
This function is an extended version of [stats::dist()] which computes
  and returns the distance matrix using the specified distance measure to compute
  the distances between the rows or columns of a data matrix.
}
