% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_median.R
\name{find_median}
\alias{find_median}
\title{Find median sample from dataset}
\usage{
find_median(x, which = c("col", "row"), na.rm = TRUE, distance = "rmsd", ...)
}
\arguments{
\item{...}{}
}
\description{
Implements a similar function to libaffy IRON findmedian function
}
\details{
This function approximates the findmedian program from libaffy. Briefly,
a medoid is identified using the rmsd distance metric and finding the
sample with the smallest all-sample mean distance. There are numerous
differences with the C code which is highly specific, but this function
roughly approximates the functionality.

Note that a more complete view of this process is given in the vignette
or in the functions [medoid()] or [find_exemplar()] in this package.
}
