% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{geometric_mean}
\alias{geometric_mean}
\title{Calculate geometric mean of vector}
\usage{
geometric_mean(x, na.rm = TRUE, positive_only = TRUE)
}
\arguments{
\item{x}{Vector}

\item{na.rm}{(TRUE) Remove NA before calculating}

\item{positive_only}{(TRUE) Are only positive values used in geometric mean?}
}
\value{
The geometric mean of the input vector
}
\description{
Calculate geometric mean of vector
}
