% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{iron_filter2}
\alias{iron_filter2}
\title{Filter observations based on signal}
\usage{
iron_filter2(x, y, min_value = 1e-05, max_value = Inf, verbose = TRUE)
}
\arguments{
\item{x}{Vector}

\item{y}{Vector}

\item{min_value}{Filter out values <= min_value (default 1e-05)}

\item{max_value}{Filter out values >= max_value (default Inf)}
}
\value{
A vector of logical values indicating whether or not *keep* values
  based on parallel comparison of x and y.
}
\description{
Filter observations in two vectors based on IRON definitions.
}
\details{
This filter implements logic associated with the IRON approach to normalization.
The logic is not completely transparent, but is included in text below:

- Exclude missing values in x or y
- Exclude values <= min_value
- Exclude values >= max_value
- Exclude values that are the minimum (>= min_value) of the vector

If these conditions result in no values being available (all filtered out),
then only the top 2 conditions are used for filtering.
}
