% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cel_utils.R
\name{list.celfiles}
\alias{list.celfiles}
\title{List CEL files in Directory/Folder}
\usage{
list.celfiles(
  path = ".",
  pattern = NULL,
  ends_with = "\\\\.[cC][eE][lL]\\\\.gz$|\\\\.[cC][eE][lL]$",
  ...
)
}
\arguments{
\item{path}{The path to list for files}

\item{pattern}{Any regex pattern to select cel file name (see [stringr::str_detect]).}

\item{ends_with}{Extension regex (e.g., .CEL)}

\item{...}{Any other parameters to pass to [fs::dir_info].}
}
\value{
A character vector of filenames
}
\description{
This function produces a character vector containing the names of files in the
named directory ending in .CEL or .cel.
}
\note{
If the directory does not exist, this function will cause an error indicating
this condition (rather than return nothing).
}
