% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exemplar.R
\name{medoid}
\alias{medoid}
\title{Find the medoid of a matrix}
\usage{
medoid(x, which = c("col", "row"), na.rm = TRUE, distance = "euclidean", ...)
}
\arguments{
\item{x}{A matrix to compare columns or rows}

\item{which}{Identify the medoid "row" or "col"}

\item{na.rm}{Should NA be removed prior to distance calculation (default TRUE)}

\item{distance}{A string representing a distance measure to use (see [distance()])}

\item{...}{Any parameters to pass to the distance function}
}
\value{
An object of type [exemplar()] representing the exemplar.
}
\description{
Find the column (row) of the matrix with the minimum summed distance
 to all other columns (rows).
}
\details{
The medoid is defined as the example with the minimum summed distances to
all other examples in the dataset. This function calculates all pair-wise distances
between columns (or rows) and then identifies the columns (or row) with the minimum
distance.

Some notes:

- If multiple columns (rows) all have a minimum summed distance, then the first column (row) with
the identical minimum distance is selected (see [base::which.min()]).

- The distance measure must be specified as a string. See [distance()] in this package
for details on available options.

- An assumption is made in this function that the distance measure is symmetric. That is,
the distance between `A` and `B` is the distance between `B` and `A`.
}
