% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{multimetric_distance}
\alias{multimetric_distance}
\title{Use multiple metrics combined as distance metric}
\usage{
multimetric_distance(x, y, methods, summarize_by = geometric_mean, ...)
}
\arguments{
\item{x}{Vector 1}

\item{y}{Vector 2}

\item{methods}{A list of methods (functions) to apply and combine via geometric mean}
}
\value{
The distance between x and y by calculating the geometric mean of
 a number of different methods.
}
\description{
Use multiple metrics combined as distance metric
}
\details{
Note that this particular approach can be very misleading, as distance
metrics (or measures) can be used and may be on different scales.
}
