% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_median.R
\name{na.dist_norm_iron}
\alias{na.dist_norm_iron}
\title{Calculate conditional probability of overlap given query}
\usage{
na.dist_norm_iron(x, y)
}
\arguments{
\item{y}{}
}
\description{
Calculate the conditional probability between two vectors as the
proportion non-NA values in the second vector by the total number of non-NA values
in common.
}
\details{
Conditional probability of #overlap given #query.

This is asymmetric, but that's what we want in this case.
We are searching for candidate reference samples.
Say that we have a reference with 100% present, and the query
is only 25% present.  We don't want to use some similarity metric
that combines those two #present together, since we want to know
if the query is maximally covered or not.  #overlap / #query will
be 1.0, since it is fully covered.  The other way around,
#overlap / #target, would be 0.25.  This is the scoring behavior
we want, since the first case indicates a good reference sample
and the 2nd case indicates a poorer reference sample.
}
