% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cel_utils.R
\name{perfect_match_probes}
\alias{perfect_match_probes}
\title{Extract perfect match (PM) probes}
\usage{
perfect_match_probes(x)
}
\arguments{
\item{x}{An AffyBatch or ExpressionFeatureSet}
}
\value{
A matrix of PM values
}
\description{
This function extracts PM probes from the object based on the object type.
}
