% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cel_utils.R
\name{read.affybatch}
\alias{read.affybatch}
\title{Read CEL files into an AffyBatch}
\usage{
read.affybatch(dir, pattern = NULL, ...)
}
\arguments{
\item{dir}{Find CEL files in dir to read}

\item{pattern}{Any CEL file pattern (regex): see [stringr::str_detect]}

\item{...}{Any other parameters to [affy::read.affybatch]}
}
\value{
An [affy::AffyBatch] object.
}
\description{
Read CEL files into an AffyBatch
}
