% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cel_utils.R
\name{read.oligo}
\alias{read.oligo}
\title{Read CEL files into an Oligo ExpressionFeatureSet}
\usage{
read.oligo(dir, pattern = NULL, ...)
}
\arguments{
\item{dir}{Find CEL files in dir to read}

\item{pattern}{Any CEL file pattern (regex): see [stringr::str_detect]}

\item{...}{Any other parameters to [oligo::read.celfiles]}
}
\value{
An [oligoClasses::ExpressionFeatureSet]
}
\description{
Read CEL files into an Oligo ExpressionFeatureSet
}
