% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{pearsonm_distance}
\alias{pearsonm_distance}
\title{Calculate distance metric based on Pearson correlation}
\usage{
pearsonm_distance(x, y, na.action = NULL)
}
\arguments{
\item{x}{Vector}

\item{y}{Vector}

\item{na.action}{Function (or NULL) to modify distance in the presence of missing
values.}
}
\description{
Calculate distance between two vectors using 1-Pearson correlation
with a correction to make a metric.
}
\details{
transform r into a metric distance [sqrt(0.5 * (1-r))] */
see Stijn van Dongen and Anton J. Enright 2012
"Metric distance derived from cosine similarity and Pearson
     *  and Spearman correlations"
}
